Move AssemblyName to shared
